requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
uvicorn==0.24.0
httpx>=0.27.0
